# backend/internal/delivery/http/openapi.yaml
openapi: 3.0.3
info:
  title: Platform RT API
  description: |
    OpenAPI 3.0 specification for the Sitransparan RT/RW backend.

    Autentikasi: `Authorization: Bearer <jwt>` (HS256, valid 24 jam).
    Role: `superadmin`, `admin_rt`, `resident`.
    Konvensi akses pada deskripsi endpoint:
    - PUBLIC: tanpa autentikasi
    - AUTH: semua role terautentikasi
    - ADMIN: superadmin + admin_rt
    - SUPERADMIN: superadmin saja

    Tenant context hanya diambil dari JWT claims yang terverifikasi (header
    X-Tenant-ID, query param, dan subdomain TIDAK dipercaya). Endpoint publik
    `/api/v1/t/{slug}/...` me-resolve tenant dari slug di path.
  version: 2.0.0
servers:
  - url: http://localhost:8081
    description: Local server (Port 8081)
  - url: /
    description: Current host relative path
paths:
...
  /api/v1/t/{slug}/rt-structure:
    get:
      summary: Bagan dan struktur pejabat kepengurusan RT resmi publik (PUBLIC)
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: '{ period: RTPeriod, members: PublicRTMember[] }'
        '404':
          description: Tenant tidak ditemukan atau nonaktif
  /api/v1/rt-structure/periods:
    get:
      summary: List periode masa bakti kepengurusan RT (AUTH)
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [draft, active, archived] } }
      responses:
        '200':
          description: '{ data: RTPeriod[] }'
    post:
      summary: Buat periode masa bakti kepengurusan RT baru (ADMIN)
      security: [{ BearerAuth: [] }]
      responses:
        '201':
          description: RTPeriod dibuat
...
  /api/v1/rt-structure/members:
    get:
      summary: List pejabat/anggota kepengurusan RT (AUTH)
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: period_id, in: query, schema: { type: string, format: uuid } }
        - { name: role, in: query, schema: { type: string } }
        - { name: section, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [aktif, demisioner] } }
      responses:
        '200':
          description: '{ data: RTMember[] }'
    post:
      summary: Tambah pejabat ke kepengurusan RT (ADMIN)
      security: [{ BearerAuth: [] }]
      responses:
        '201':
          description: RTMember ditambahkan
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...